ConnectionEntity and SalesforceId transformers tested. Outgoing compilation and queueing tested.

+++ Tests/Salesforce/SalesforceConnectorTest.php
@@ -8,17 +8,18 @@
 
 namespace AE\ConnectBundle\Tests\Salesforce;
 
-use AE\ConnectBundle\Driver\DbalConnectionDriver;
 use AE\ConnectBundle\Salesforce\Inbound\SalesforceConsumerInterface;
 use AE\ConnectBundle\Salesforce\Outbound\Enqueue\OutboundProcessor;
 use AE\ConnectBundle\Salesforce\Outbound\Queue\OutboundQueue;
 use AE\ConnectBundle\Salesforce\SalesforceConnector;
 use AE\ConnectBundle\Tests\DatabaseTestCase;
 use AE\ConnectBundle\Tests\Entity\Account;
+use AE\ConnectBundle\Tests\Entity\ConnectionEntity;
 use AE\ConnectBundle\Tests\Entity\Contact;
 use AE\ConnectBundle\Tests\Entity\Order;
 use AE\ConnectBundle\Tests\Entity\Product;
 use AE\SalesforceRestSdk\Model\SObject;
+use Doctrine\ORM\EntityRepository;
 use Enqueue\Client\DriverInterface;
 use Enqueue\Consumption\Result;
 use Enqueue\Fs\FsContext;
@@ -54,6 +55,12 @@ class SalesforceConnectorTest extends DatabaseTestCase
         $this->connector->disable();
         $this->loadOrgConnections();
 
+        $this->loadFixtures(
+            [
+                $this->getProjectDir().'/Tests/Resources/config/connections.yml',
+            ]
+        );
+
         $manager  = $this->doctrine->getManager();
         $items    = [];
         $queue    = $this->driver->createQueue('default');
@@ -61,22 +68,14 @@ class SalesforceConnectorTest extends DatabaseTestCase
         /** @var OutboundProcessor $processor */
         $processor = $this->get(OutboundProcessor::class);
 
-        $this->context->purge($queue);
+        if (method_exists($this->context, 'purge')) {
+            $this->context->purge($queue);
+        }
 
         for ($i = 0; $i < 5; $i++) {
             $this->createOrder($items);
             $manager->flush();
         }
-
-        // Create one for the non-default connection
-        $account = new Account();
-        $account->setName('Test DB Account');
-        $account->setConnection('db_test_org1');
-        $manager->persist($account);
-
-        $items[] = $account;
-
-        $manager->flush();
 
         $this->connector->enable();
 
@@ -84,6 +83,64 @@ class SalesforceConnectorTest extends DatabaseTestCase
             $this->connector->send($item);
         }
 
+        while (null !== ($message = $consumer->receive(100))) {
+            switch ($processor->process($message, $this->context)) {
+                case Result::ACK:
+                    $consumer->acknowledge($message);
+                    break;
+                case Result::REJECT:
+                    $consumer->reject($message, false);
+                    break;
+                case Result::REQUEUE:
+                    $consumer->reject($message, true);
+                    break;
+            }
+        }
+
+        $this->get(OutboundQueue::class)->send();
+
+        $accounts = $manager->getRepository(Account::class)->findBy(['sfid' => null, 'connection' => 'default']);
+        $this->assertEmpty($accounts);
+
+        $orders = $manager->getRepository(Order::class)->findBy(['sfid' => null]);
+        $this->assertEmpty($orders);
+
+        $this->assertCount(5, $manager->getRepository(Account::class)->findBy(['connection' => 'default']));
+    }
+
+    public function testOutgoingDbTest()
+    {
+        // We don't want to fire on any triggers now do we?
+        $this->connector->disable();
+        $this->loadOrgConnections();
+
+        $this->loadFixtures(
+            [
+                $this->getProjectDir().'/Tests/Resources/config/connections.yml',
+            ]
+        );
+
+        $manager  = $this->doctrine->getManager();
+        $queue    = $this->driver->createQueue('default');
+        $consumer = $this->context->createConsumer($queue);
+        /** @var OutboundProcessor $processor */
+        $processor = $this->get(OutboundProcessor::class);
+        /** @var ConnectionEntity $conn */
+        $conn      = $manager->getRepository(ConnectionEntity::class)->findOneBy(['name' => 'db_test_org1']);
+
+        if (method_exists($this->context, 'purge')) {
+            $this->context->purge($queue);
+        }
+
+        // Create one for the non-default connection
+        $account = new Account();
+        $account->setName('Test DB Account');
+        $account->setConnections([$conn]);
+        $manager->persist($account);
+
+        $manager->flush();
+
+        $this->connector->enable();
         $this->connector->send($account, 'db_test_org1');
 
         while (null !== ($message = $consumer->receive(100))) {
@@ -102,14 +159,18 @@ class SalesforceConnectorTest extends DatabaseTestCase
 
         $this->get(OutboundQueue::class)->send();
 
-        $accounts = $manager->getRepository(Account::class)->findBy(['sfid' => null]);
-        $this->assertEmpty($accounts);
+        /** @var EntityRepository $repo */
+        $repo = $manager->getRepository(Account::class);
+        $qb = $repo->createQueryBuilder('a');
 
-        $orders = $manager->getRepository(Order::class)->findBy(['sfid' => null]);
-        $this->assertEmpty($orders);
+        $qb->join('a.connections', 'c')
+            ->where('c.id = :conn')
+            ->setParameter('conn', $conn->getId())
+            ;
 
-        $this->assertCount(1, $manager->getRepository(Account::class)->findBy(['connection' => 'db_test_org1']));
-        $this->assertCount(5, $manager->getRepository(Account::class)->findBy(['connection' => 'default']));
+        $accounts = $qb->getQuery()->getResult();
+
+        $this->assertCount(1, $accounts);
     }
 
     public function testIncoming()
